#172: RegisterCategorySelectWithRankTableViewCell mainView로 RegisterCategorySelectView 설정 > RegisterCategorySelectTableViewCell 상속 제거
